Initial support for plugin control (enable/disable, options, etc)  also consolidated common plugin code to PluginManagerBase.cs and made all plugins inherit from IPlugin.

+++ IHeadsetPlugin.cs
@@ -46,26 +46,32 @@ namespace FSClient {
 		}
 	}
 
-	public class HeadsetPluginManager : IDisposable {
+	public class HeadsetPluginManager : PluginManagerBase {
 		private class DeviceData {
 			public IDeviceHost host;
 			public IHeadsetDevice device;
 			public bool active;
 			public bool in_use;
-			public PluginData plugin;
+			public HeadsetPluginData plugin;
 			public IHeadsetDevice.HEADSET_LIMITATIONS limitations;
 			public bool HasLimit(IHeadsetDevice.HEADSET_LIMITATIONS limit) {
 				return (limitations & limit) == limit;
 			}
 		}
-		private class PluginData{
-			public IHeadsetPlugin plugin;
+		private class HeadsetPluginData : PluginData {
+			public HeadsetPluginData(PluginData data) : base(data){}
+			public HeadsetPluginData() : base() { }
+
+			public IHeadsetPlugin headset_plugin {
+				get { return plugin as IHeadsetPlugin; }
+				set { plugin = value; }
+			}
 			public int error_tries_left = 0;
 			public DateTime last_error_time = DateTime.Now;//start with now so error count will not reset right away
 		}
-		private class PluginError{
+		private class PluginError {
 			public DeviceData device;
-			public PluginData plugin;
+			public HeadsetPluginData plugin;
 			public Exception exception;
 		}
 		private object devices_lock = new object();
@@ -73,7 +79,23 @@ namespace FSClient {
 		private const int DEFAULT_PLUGIN_RETRIES = 5;
 		private List<DeviceData> devices = new List<DeviceData>();
 
-		private List<PluginData> plugins = new List<PluginData>();
+		private List<HeadsetPluginData> plugins = new List<HeadsetPluginData>();
+		public static HeadsetPluginManager GetPluginManager(SettingsPluginDataCollection settings) {
+			HeadsetPluginManager manager = new HeadsetPluginManager();
+			manager.LoadSettings(settings);
+			return manager;
+		}
+		public override IEnumerable<PluginData> GetPlugins() {
+			return plugins;
+		}
+		protected override void SetPlugins(IEnumerable<PluginData> plugins) {
+			if (this.plugins.Count > 0)
+				throw new Exception("Cannot call SetPlugins if plugins have already been loaded");
+			foreach (PluginData data in plugins)
+				this.plugins.Add(new HeadsetPluginData(data));
+
+		}
+
 		public string[] AvailableDevices() {
 			return (from d in devices select d.device.GetName()).ToArray();
 		}
@@ -105,7 +127,7 @@ namespace FSClient {
 					catch (Exception e) {
 						if (errors == null)
 							errors = new List<PluginError>();
-						errors.Add(new PluginError{device = device, exception = e});
+						errors.Add(new PluginError { device = device, exception = e });
 					}
 				}
 			}
@@ -135,14 +157,14 @@ namespace FSClient {
 			List<PluginError> errors = null;
 
 			lock (devices_lock) {
-				foreach (DeviceData data in devices){
+				foreach (DeviceData data in devices) {
 					if (!data.active)
 						continue;
-					try{
+					try {
 						if (type != IDeviceHost.PHONE_EVENT_TYPE.LineActive)
 							data.host.CreateEvent(type, enable);
-						else{
-							if (data.HasLimit(IHeadsetDevice.HEADSET_LIMITATIONS.ONLY_ACTIVE_DURING_CALLS)){
+						else {
+							if (data.HasLimit(IHeadsetDevice.HEADSET_LIMITATIONS.ONLY_ACTIVE_DURING_CALLS)) {
 								if (is_during_call)
 									data.host.CreateEvent(type, enable);
 							}
@@ -150,7 +172,7 @@ namespace FSClient {
 								data.host.CreateEvent(type, enable);
 						}
 					}
-					catch (Exception e){
+					catch (Exception e) {
 						if (errors == null)
 							errors = new List<PluginError>();
 						errors.Add(new PluginError { device = data, exception = e });
@@ -222,7 +244,7 @@ namespace FSClient {
 			}
 
 		}
-		private void DeviceAdded(object sender, IHeadsetPlugin.DeviceEventArgs e){
+		private void DeviceAdded(object sender, IHeadsetPlugin.DeviceEventArgs e) {
 
 			IHeadsetPlugin plugin = sender as IHeadsetPlugin;
 			DeviceData data = new DeviceData { device = e.device, host = new IDeviceHost(), active = false, in_use = false };
@@ -240,10 +262,10 @@ namespace FSClient {
 				}
 			}
 			catch (Exception exp) {
-				HandleError(new PluginError{device = data, exception = exp});
+				HandleError(new PluginError { device = data, exception = exp });
 			}
 		}
-		
+
 		private void DeviceRemoved(object sender, IHeadsetPlugin.DeviceEventArgs e) {
 			DeviceData data = (from d in devices where d.device == e.device select d).SingleOrDefault();
 			if (data == null)
@@ -256,28 +278,6 @@ namespace FSClient {
 				active_device_name = null;
 			}
 		}
-		public void RegisterPlugin(IHeadsetPlugin plugin) {
-			PluginData data = new PluginData() { plugin = plugin };
-			plugins.Add(data);
-			try{
-				plugin.DeviceAdded += DeviceAdded;
-				plugin.DeviceRemoved += DeviceRemoved;
-				plugin.Initialize();
-				data.error_tries_left = DEFAULT_PLUGIN_RETRIES;
-			}
-			catch (Exception e) {
-				HandleError(new PluginError { plugin = data, exception = e });
-			}
-		}
-		private bool IsTypeOf(Type to_check, Type of) {
-			if (to_check == null)
-				return false;
-			if (to_check == of)
-				return true;
-			return IsTypeOf(to_check.BaseType, of);
-		}
-
-
 
 		private void BrokerDevicesReadyChanged(object sender, bool data) {
 			LineActive(data, false);
@@ -294,55 +294,66 @@ namespace FSClient {
 				SetCallerID(Call.active_call.other_party_name, Call.active_call.other_party_number);
 			Ring(data);
 		}
-		public HeadsetPluginManager() {
+		public HeadsetPluginManager(){
 			broker = Broker.get_instance();
-			broker.DevicesReadyChanged += BrokerDevicesReadyChanged;
-			broker.call_answeredChanged += BrokerAnsweredChanged;
-			broker.active_call_ringingChanged += BrokerCallRingingChanged;
-			broker.MutedChanged += BrokerMuteChanged;
-			string plugin_dir = Utils.plugins_dir();
-			string[] dlls;
-			try{
-				dlls = Directory.GetFileSystemEntries(plugin_dir, "*.dll");
-			}catch (DirectoryNotFoundException){
-				return;
-			}
-			foreach (String dll in dlls) {
-				try{
-					Assembly asm = Assembly.LoadFrom(dll);
-					if (asm == null)
-						continue;
-					foreach (Type type in asm.GetTypes()) {
-						if (type.IsAbstract)
-							continue;
-						if (!IsTypeOf(type, typeof(IHeadsetPlugin)))
-							continue;
+		}
 
-						RegisterPlugin(asm.CreateInstance(type.FullName) as IHeadsetPlugin);
-					}
-				}
-				catch (ReflectionTypeLoadException ex){
-					String err = "Error creating headset plugin from dll \"" + dll + "\" due to a loader exception, make sure you have the headset runtime/api/sdk installed error was:\n";
-					foreach (Exception e in ex.LoaderExceptions)
-						err += e.Message + "\n";
-					Utils.PluginLog("Headset Plugin Manager",err);
-				}
+		public override void LoadPlugins(){
+			LoadActualPlugins(typeof (IHeadsetPlugin),  plugins);
 
-				catch (Exception e){
-					Utils.PluginLog("Headset Plugin Manager", "Error creating headset plugin from dll \"" + dll + "\" of: " + e.Message);
-				}
+			if ((from p in plugins where p.enabled && p.state != PluginData.PluginDataState.ERROR_LOADING select true).Count() > 0){
+				broker.DevicesReadyChanged += BrokerDevicesReadyChanged;
+				broker.call_answeredChanged += BrokerAnsweredChanged;
+				broker.active_call_ringingChanged += BrokerCallRingingChanged;
+				broker.MutedChanged += BrokerMuteChanged;
 			}
 		}
+		protected override void HandlePluginLoadReflectionException(PluginData data, ReflectionTypeLoadException ex) {
+			String err = "Error creating headset plugin from dll \"" + data.dll + "\" due to a loader exception, make sure you have the headset runtime/api/sdk installed error was:\n";
+			foreach (Exception e in ex.LoaderExceptions)
+				err += e.Message + "\n";
+			data.last_error = err;
+			data.state = PluginData.PluginDataState.ERROR_LOADING;
+			Utils.PluginLog(PluginManagerName(), err);
+		}
+		public override string PluginManagerName(){
+			return "Headset Plugin Manager";
+		}
+
+		protected override void PluginLoadAddPlugin(PluginData plugin){
+			plugins.Add(plugin as HeadsetPluginData);
+		}
+
+		protected override void PluginLoadRegisterPlugin(PluginData plugin){
+			HeadsetPluginData data = plugin as HeadsetPluginData;
+
+			try {
+				data.headset_plugin.DeviceAdded += DeviceAdded;
+				data.headset_plugin.DeviceRemoved += DeviceRemoved;
+				data.headset_plugin.Initialize();
+				data.error_tries_left = DEFAULT_PLUGIN_RETRIES;
+				data.state = PluginData.PluginDataState.LOADED;
+			}
+			catch (Exception e) {
+				HandleError(new PluginError { plugin = data, exception = e });
+			}
+		}
+
+		protected override PluginData NewPluginData(String dll){
+			return new HeadsetPluginData();
+		}
+
 		~HeadsetPluginManager() {
 			Dispose();
 		}
 		private bool disposed;
-		public void Dispose() {
+		public override void Dispose() {
 			if (!disposed) {
 				disposed = true;
-				foreach (PluginData plugin_data in plugins) {
+				foreach (HeadsetPluginData plugin_data in plugins) {
 					try {
-						plugin_data.plugin.Terminate();
+						if (plugin_data.plugin != null)
+							plugin_data.plugin.Terminate();
 					}
 					catch (Exception e) { Utils.PluginLog("Headset Plugin Manager", "Error terminating a plugin: " + e.Message); }
 				}
@@ -358,16 +369,16 @@ namespace FSClient {
 
 			error.plugin.last_error_time = DateTime.Now;
 			String restart_msg = error.plugin.error_tries_left > 0 ? " will try to restart/init it " + error.plugin.error_tries_left + " more times" : " will not be restarting it";
-			Utils.PluginLog("Headset Plugin Manager", "Plugin " + error.plugin.plugin.ProviderName() + " had an error Due to: " + error.exception.Message + "\n" + restart_msg);
+			Utils.PluginLog(PluginManagerName(), "Plugin " + error.plugin.plugin.ProviderName() + " had an error Due to: " + error.exception.Message + "\n" + restart_msg);
 			List<DeviceData> to_remove = new List<DeviceData>();
 			lock (devices_lock) {
-				foreach (DeviceData device in devices){
+				foreach (DeviceData device in devices) {
 					if (device.plugin == error.plugin) {
-						try{
+						try {
 							device.device.SetActive(false);
 						}
-						catch (Exception){
-							Utils.PluginLog("Headset Plugin Manager", "While handling error wasn't able to deactivate device, not a major issue");
+						catch (Exception) {
+							Utils.PluginLog(PluginManagerName(), "While handling error wasn't able to deactivate device, not a major issue");
 						}
 						to_remove.Add(device);
 					}
@@ -381,8 +392,8 @@ namespace FSClient {
 			if (error.plugin.error_tries_left-- > 0)
 				DelayedFunction.DelayedCall("IHeadsetPlugin_PluginStart_ " + error.plugin.plugin.ProviderName(), () => init_plugin(error.plugin), 1000);//give it a second
 		}
-		private void init_plugin(PluginData plugin){
-			try{
+		private void init_plugin(HeadsetPluginData plugin) {
+			try {
 				plugin.plugin.Initialize();
 			}
 			catch (Exception e) {
@@ -391,7 +402,7 @@ namespace FSClient {
 		}
 	}
 
-	public abstract class IHeadsetPlugin {
+	public abstract class IHeadsetPlugin : IPlugin {
 		public class DeviceEventArgs : EventArgs {
 			public readonly IHeadsetDevice device;
 			public DeviceEventArgs(IHeadsetDevice device) {
@@ -400,8 +411,5 @@ namespace FSClient {
 		}
 		public EventHandler<DeviceEventArgs> DeviceAdded;
 		public EventHandler<DeviceEventArgs> DeviceRemoved;
-		public abstract string ProviderName();
-		public abstract void Initialize();
-		public abstract void Terminate();
 	}
 }
